Singly LinkedList
The above code singular LinkedList which consists of various . 
Append method to add element at last position.
prepend method to add element at first position.
Insert method to add a element at specific position.
toString method to display the LinkedList.
Search method to find the index value of element in LinkedList , if element not in LinkedList return -1 to indicate element is not in LinkedList.
Pop method to remove the element from last index. , if LinkedList is empty return -1 to indicate LinkedList is empty.